Moved Railtie code into its own class file so that the gem can be used in non Rails apps
